requirements.txt
pandas
pytest
python-dotenv
pytest-cov
pylint
logging